requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
anthropic>=0.18.0
trafilatura>=1.6.0
gspread>=5.12.0
google-auth>=2.25.0
google-auth-oauthlib>=1.2.0
